fix(listeners): correctly unwatch multiple queries to the same base path (#235)
* Correctly unset multiple queries to the same base path

Passing queryId allows into the unWatchEvent function is necessary for
unique queries to the same base path to be properly identified when
unwatching events..  getQueryIdFromPath and getWatchPath do not
correctly identify unique queries at the same path (because the path
that is passed into these functions is the path without query strings
attached).  This scenario happens when using storeAs functionality.